% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship_function.R
\name{travel_ship_description}
\alias{travel_ship_description}
\title{Get the ship that share the same ship type}
\usage{
travel_ship_description(ship_subset)
}
\arguments{
\item{ship_data}{Data set with the ship travel information}

\item{type_of_ship}{Type of the ship we are interested in}
}
\value{
`Vector` with the names that are of the selected type
}
\description{
Get the ship that share the same ship type
}
\examples{

filter_ship_type(ship_data = ship_data,type_of_ship = "Cargo")

}
